models/PPO/demo/requirements.txt
flask==2.3.3
torch
sentence-transformers
scikit-learn
nltk
openai
python-dotenv
requests 